Attempt to fix GH Actions Windows timeouts
