+++ src/Bicep.Core/TypeSystem/EmptyResourceTypeProvider.cs
@@ -19,5 +19,7 @@ namespace Bicep.Core.TypeSystem
 
         public bool HasDefinedType(ResourceTypeReference typeReference)
             => false;
+
+        public void ClearCaches() { }
     }
 }
